+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 1.4.74 (unreleased)
 ------------------------
+* is_manager replaced by can_edit_users, bound to permission, not role
+  [dumitval]
 
 1.4.73 (2015-03-30)
 ------------------------